feat(duckdb): integrate storage backend, CloudEvents, and address review feedback
