+++ Cargo.toml
@@ -141,3 +141,4 @@ debug = "line-tables-only"
 # Optimize dependency builds to improve iterative development
 [profile.dev.package."*"]
 opt-level = 3
+debug = false
