+++ memory-storage-duckdb/src/storage/monitoring.rs
@@ -11,9 +11,8 @@ impl DuckDbStorage {
     ) -> Result<()> {
         let conn_arc = Arc::clone(&self.conn);
         let record = record.clone();
-        let duration_ms = i64::try_from(record.duration.as_millis()).map_err(|e| {
-            Error::Storage(format!("Duration overflow for execution record: {e}"))
-        })?;
+        let duration_ms = i64::try_from(record.duration.as_millis())
+            .map_err(|e| Error::Storage(format!("Duration overflow for execution record: {e}")))?;
 
         tokio::task::spawn_blocking(move || {
             let conn = conn_arc.lock();
@@ -46,30 +45,22 @@ impl DuckDbStorage {
         let conn_arc = Arc::clone(&self.conn);
         let metrics = metrics.clone();
 
-        let total_executions = i64::try_from(metrics.total_executions).map_err(|e| {
-            Error::Storage(format!("total_executions overflow: {e}"))
-        })?;
-        let successful_executions = i64::try_from(metrics.successful_executions).map_err(|e| {
-            Error::Storage(format!("successful_executions overflow: {e}"))
-        })?;
-        let total_duration_ms = i64::try_from(metrics.total_duration.as_millis()).map_err(|e| {
-            Error::Storage(format!("total_duration overflow: {e}"))
-        })?;
-        let avg_duration_ms = i64::try_from(metrics.avg_duration.as_millis()).map_err(|e| {
-            Error::Storage(format!("avg_duration overflow: {e}"))
-        })?;
-        let min_duration_ms = i64::try_from(metrics.min_duration.as_millis()).map_err(|e| {
-            Error::Storage(format!("min_duration overflow: {e}"))
-        })?;
-        let max_duration_ms = i64::try_from(metrics.max_duration.as_millis()).map_err(|e| {
-            Error::Storage(format!("max_duration overflow: {e}"))
-        })?;
-        let current_streak = i32::try_from(metrics.current_streak).map_err(|e| {
-            Error::Storage(format!("current_streak overflow: {e}"))
-        })?;
-        let longest_streak = i32::try_from(metrics.longest_streak).map_err(|e| {
-            Error::Storage(format!("longest_streak overflow: {e}"))
-        })?;
+        let total_executions = i64::try_from(metrics.total_executions)
+            .map_err(|e| Error::Storage(format!("total_executions overflow: {e}")))?;
+        let successful_executions = i64::try_from(metrics.successful_executions)
+            .map_err(|e| Error::Storage(format!("successful_executions overflow: {e}")))?;
+        let total_duration_ms = i64::try_from(metrics.total_duration.as_millis())
+            .map_err(|e| Error::Storage(format!("total_duration overflow: {e}")))?;
+        let avg_duration_ms = i64::try_from(metrics.avg_duration.as_millis())
+            .map_err(|e| Error::Storage(format!("avg_duration overflow: {e}")))?;
+        let min_duration_ms = i64::try_from(metrics.min_duration.as_millis())
+            .map_err(|e| Error::Storage(format!("min_duration overflow: {e}")))?;
+        let max_duration_ms = i64::try_from(metrics.max_duration.as_millis())
+            .map_err(|e| Error::Storage(format!("max_duration overflow: {e}")))?;
+        let current_streak = i32::try_from(metrics.current_streak)
+            .map_err(|e| Error::Storage(format!("current_streak overflow: {e}")))?;
+        let longest_streak = i32::try_from(metrics.longest_streak)
+            .map_err(|e| Error::Storage(format!("longest_streak overflow: {e}")))?;
 
         tokio::task::spawn_blocking(move || {
             let conn = conn_arc.lock();
@@ -108,15 +99,12 @@ impl DuckDbStorage {
         let conn_arc = Arc::clone(&self.conn);
         let metrics = metrics.clone();
 
-        let total_tasks = i64::try_from(metrics.total_tasks).map_err(|e| {
-            Error::Storage(format!("total_tasks overflow: {e}"))
-        })?;
-        let completed_tasks = i64::try_from(metrics.completed_tasks).map_err(|e| {
-            Error::Storage(format!("completed_tasks overflow: {e}"))
-        })?;
-        let avg_completion_time_ms = i64::try_from(metrics.avg_completion_time.as_millis()).map_err(|e| {
-            Error::Storage(format!("avg_completion_time overflow: {e}"))
-        })?;
+        let total_tasks = i64::try_from(metrics.total_tasks)
+            .map_err(|e| Error::Storage(format!("total_tasks overflow: {e}")))?;
+        let completed_tasks = i64::try_from(metrics.completed_tasks)
+            .map_err(|e| Error::Storage(format!("completed_tasks overflow: {e}")))?;
+        let avg_completion_time_ms = i64::try_from(metrics.avg_completion_time.as_millis())
+            .map_err(|e| Error::Storage(format!("avg_completion_time overflow: {e}")))?;
 
         tokio::task::spawn_blocking(move || {
             let conn = conn_arc.lock();
@@ -189,23 +177,32 @@ impl DuckDbStorage {
                 let metrics = do_memory_core::monitoring::types::AgentMetrics {
                     agent_name,
                     agent_type,
-                    total_executions: u64::try_from(total_executions).map_err(|e| Error::Storage(format!("total_executions conversion: {e}")))?,
-                    successful_executions: u64::try_from(successful_executions).map_err(|e| Error::Storage(format!("successful_executions conversion: {e}")))?,
+                    total_executions: u64::try_from(total_executions)
+                        .map_err(|e| Error::Storage(format!("total_executions conversion: {e}")))?,
+                    successful_executions: u64::try_from(successful_executions).map_err(|e| {
+                        Error::Storage(format!("successful_executions conversion: {e}"))
+                    })?,
                     total_duration: std::time::Duration::from_millis(
-                        u64::try_from(total_duration_ms).map_err(|e| Error::Storage(format!("total_duration conversion: {e}")))?
+                        u64::try_from(total_duration_ms).map_err(|e| {
+                            Error::Storage(format!("total_duration conversion: {e}"))
+                        })?,
                     ),
                     avg_duration: if total_executions > 0 {
                         std::time::Duration::from_millis(
-                            u64::try_from(total_duration_ms / total_executions).map_err(|e| Error::Storage(format!("avg_duration conversion: {e}")))?
+                            u64::try_from(total_duration_ms / total_executions).map_err(|e| {
+                                Error::Storage(format!("avg_duration conversion: {e}"))
+                            })?,
                         )
                     } else {
                         std::time::Duration::ZERO
                     },
                     min_duration: std::time::Duration::from_millis(
-                        u64::try_from(min_duration_ms).map_err(|e| Error::Storage(format!("min_duration conversion: {e}")))?
+                        u64::try_from(min_duration_ms)
+                            .map_err(|e| Error::Storage(format!("min_duration conversion: {e}")))?,
                     ),
                     max_duration: std::time::Duration::from_millis(
-                        u64::try_from(max_duration_ms).map_err(|e| Error::Storage(format!("max_duration conversion: {e}")))?
+                        u64::try_from(max_duration_ms)
+                            .map_err(|e| Error::Storage(format!("max_duration conversion: {e}")))?,
                     ),
                     last_execution: last_execution_str
                         .map(|s| {
@@ -215,8 +212,10 @@ impl DuckDbStorage {
                                 .map_err(|e| Error::Storage(format!("last_execution parse: {e}")))
                         })
                         .transpose()?,
-                    current_streak: u32::try_from(current_streak).map_err(|e| Error::Storage(format!("current_streak conversion: {e}")))?,
-                    longest_streak: u32::try_from(longest_streak).map_err(|e| Error::Storage(format!("longest_streak conversion: {e}")))?,
+                    current_streak: u32::try_from(current_streak)
+                        .map_err(|e| Error::Storage(format!("current_streak conversion: {e}")))?,
+                    longest_streak: u32::try_from(longest_streak)
+                        .map_err(|e| Error::Storage(format!("longest_streak conversion: {e}")))?,
                 };
                 Ok::<Option<do_memory_core::monitoring::types::AgentMetrics>, Error>(Some(metrics))
             } else {
@@ -286,7 +285,8 @@ impl DuckDbStorage {
                     agent_type,
                     success,
                     duration: std::time::Duration::from_millis(
-                        u64::try_from(duration_ms).map_err(|e| Error::Storage(format!("duration conversion: {e}")))?
+                        u64::try_from(duration_ms)
+                            .map_err(|e| Error::Storage(format!("duration conversion: {e}")))?,
                     ),
                     started_at,
                     task_description,
@@ -331,10 +331,14 @@ impl DuckDbStorage {
 
                 let metrics = do_memory_core::monitoring::types::TaskMetrics {
                     task_type,
-                    total_tasks: u64::try_from(total_tasks).map_err(|e| Error::Storage(format!("total_tasks conversion: {e}")))?,
-                    completed_tasks: u64::try_from(completed_tasks).map_err(|e| Error::Storage(format!("completed_tasks conversion: {e}")))?,
+                    total_tasks: u64::try_from(total_tasks)
+                        .map_err(|e| Error::Storage(format!("total_tasks conversion: {e}")))?,
+                    completed_tasks: u64::try_from(completed_tasks)
+                        .map_err(|e| Error::Storage(format!("completed_tasks conversion: {e}")))?,
                     avg_completion_time: std::time::Duration::from_millis(
-                        u64::try_from(avg_completion_time_ms).map_err(|e| Error::Storage(format!("avg_completion_time conversion: {e}")))?
+                        u64::try_from(avg_completion_time_ms).map_err(|e| {
+                            Error::Storage(format!("avg_completion_time conversion: {e}"))
+                        })?,
                     ),
                     agent_success_rates: serde_json::from_str(&agent_success_rates_json)
                         .map_err(|e| Error::Storage(e.to_string()))?,
