+++ plans/GOAP_CLOUDEVENTS_INTEGRATION.md
@@ -1,15 +1,18 @@
 # GOAP: CloudEvents Integration
 
 ## Goal
+
 Standardize agent lifecycle events using CNCF CloudEvents.
 
 ## Current State
+
 - New crate `do-memory-events` integrated.
 - `EventEmitter` trait implemented and wired.
 - Basic lifecycle events instrumented.
 - Standardized to `f64` scores.
 
 ## Tasks
+
 - [x] Create `do-memory-events` crate.
 - [x] Define `MemoryEvent` variants.
 - [x] Implement CloudEvents mapping.
