+++ plans/GOAP_CLOUDEVENTS_STABILIZATION.md
@@ -1,9 +1,11 @@
 # GOAP: CloudEvents Stabilization and Expansion
 
 ## Goal
+
 Harden the CloudEvents implementation and expand its reach across the workspace.
 
 ## Tasks
+
 - [ ] Implement retry logic for `HttpEventEmitter`.
 - [ ] Add WebSocket event sink to `do-memory-events`.
 - [ ] Integrate CloudEvents forwarding into `do-memory-mcp` tool calls.
